% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_access.R
\name{access_meteo}
\alias{access_meteo}
\title{Download and preprocess meteorological data from three station networks}
\usage{
access_meteo(
  networks,
  datetime_utc_obs,
  lon_obs,
  lat_obs,
  deg_filter = 2,
  time_thresh_s = 3600,
  dist_thresh_m = 1e+05
)
}
\arguments{
\item{networks}{One of "ALL", "HADS", "LCD", or "WCC"}

\item{datetime_utc_obs}{POSIX-formatted UTC datetime for which data should be gathered}

\item{lon_obs}{Longitude in decimal degrees}

\item{lat_obs}{Latitude in decimal degrees}

\item{deg_filter}{Number of degrees surrounding the point location to which the station search should be limited}

\item{time_thresh_s}{Time (in seconds) to be added to/subtracted from datetime_utc_obs, thus defining the time extent for the data search}

\item{dist_thresh_m}{Distance (in meters) that a station must be within to be considered}
}
\value{
a dataframe of meteorological data
}
\description{
Download and preprocess meteorological data from three station networks
}
\examples{
\dontrun{
met_network = "HADS"
datetime = as.POSIXct("2023-01-01 16:00:00", tz = "UTC")
lon = -105
lat = 40
degree_filter = 1
meteo <- access_meteo(networks = met_network,
                      datetime_utc_obs = datetime,
                      lon_obs = lon,
                      lat_obs = lat,
                      deg_filter = degree_filter)
}
}
