% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{construct_gpm_base_url}
\alias{construct_gpm_base_url}
\title{Construct the base URL for GPM IMERG data for a given date timestamp}
\usage{
construct_gpm_base_url(date_of_interest)
}
\arguments{
\item{date_of_interest}{character string date in YYYY-MM-DDTHH:MM:SS.000Z format UTC time (e.g. "2024-01-26T17:16:43.000Z")}
}
\value{
character GPM IMERG base URL for the given date
}
\description{
Construct the base URL for GPM IMERG data for a given date timestamp
}
