% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_access.R
\name{download_meteo_madis}
\alias{download_meteo_madis}
\title{Download meteorological data from MADIS}
\usage{
download_meteo_madis(lon_obs, lat_obs, deg_filter, datetime_utc_obs)
}
\arguments{
\item{lon_obs}{Start of search window as POSIX-formatted UTC datetime}

\item{lat_obs}{End of search window as POSIX-formatted UTC datetime}

\item{deg_filter}{Search radius in degrees}

\item{datetime_utc_obs}{dataframe of station metadata (from station_select)}
}
\value{
List of meteorological data and station metadata
}
\description{
Download meteorological data from MADIS
}
