% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_convert.R
\name{f_to_c}
\alias{f_to_c}
\title{Function to convert temperature in Fahrenheit to Celsius}
\usage{
f_to_c(temp_F)
}
\arguments{
\item{temp_F}{The temperature in degrees Fahrenheit}
}
\value{
The temperature in degrees Celsius
}
\description{
Function to convert temperature in Fahrenheit to Celsius
}
\examples{
tair_c <- f_to_c(temp_F = 32);
}
