% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_gather.R
\name{gather_meta}
\alias{gather_meta}
\title{Function to gather metadata from a list of stations}
\usage{
gather_meta(stations, network)
}
\arguments{
\item{stations}{Vector of station ids (HADS, LCD, WCC) or the station data itself for MADIS}

\item{network}{The station's network ("HADS", "LCD", "WCC", "MADIS")}
}
\value{
A dataframe of filtered station metadata
}
\description{
Function to gather metadata from a list of stations
}
