% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_access.R
\name{get_available_networks}
\alias{get_available_networks}
\title{Determine available networks based on datetime}
\usage{
get_available_networks(datetime_utc_obs, include_deprecated = FALSE)
}
\arguments{
\item{datetime_utc_obs}{POSIX-formatted UTC datetime}

\item{include_deprecated}{Logical, default FALSE. If TRUE, includes networks
that are deprecated but still functional for historical data.}
}
\value{
Character vector of available network names
}
\description{
Determine available networks based on datetime
}
\keyword{internal}
