% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_closest_url}
\alias{get_closest_url}
\title{Get the closest URL to the "match_string" string
Uses the Levenshtein distance to find the closest URL to the "match_string", match string is the "product name" string determined by the construct_gpm_product() function}
\usage{
get_closest_url(urls, match_string)
}
\arguments{
\item{urls}{character vector of URLs}

\item{match_string}{string to partial match with}
}
\value{
character of the closest URL to the "match_string"
}
\description{
Get the closest URL to the "match_string" string
Uses the Levenshtein distance to find the closest URL to the "match_string", match string is the "product name" string determined by the construct_gpm_product() function
}
\seealso{
Other imerg: 
\code{\link{construct_gpm_base_url}()},
\code{\link{construct_gpm_product}()},
\code{\link{get_final_urls}()},
\code{\link{get_imerg}()}
}
\concept{imerg}
