% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_eco_level4}
\alias{get_eco_level4}
\title{Geolocate location to assign ecoregion 4 association}
\usage{
get_eco_level4(lon_obs, lat_obs)
}
\arguments{
\item{lon_obs}{Longitude of observation in decimal degrees (°)}

\item{lat_obs}{Latitude of observation in decimal degrees (°)}
}
\value{
Ecoregion level 4
}
\description{
Geolocate location to assign ecoregion 4 association
}
\examples{
\dontrun{
lon = -105
lat = 40
ecoregion4 <- get_eco_level4(lon, lat)
}
}
