% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_elev}
\alias{get_elev}
\title{Get elevation based on lat/lon}
\usage{
get_elev(lon_obs, lat_obs)
}
\arguments{
\item{lon_obs}{Longitude of observation in decimal degrees (°)}

\item{lat_obs}{Latitude of observation in decimal degrees (°)}
}
\value{
Elevation based on location
}
\description{
Get elevation based on lat/lon
}
