% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_final_urls}
\alias{get_final_urls}
\title{Get the resource URLs from NASA GPM XML Catalog
This function reads the catalog.xml file from the hyrax GPM IMERG catalog.xml URL and
extracts the "dap" service files, and returns them as a character vector}
\usage{
get_final_urls(url, verbose = FALSE)
}
\arguments{
\item{url}{character string URL to the catalog.xml file}

\item{verbose}{logical, whether to print out messages or not}
}
\value{
vector of URLs for daily GPM IMERG resources
}
\description{
Get the resource URLs from NASA GPM XML Catalog
This function reads the catalog.xml file from the hyrax GPM IMERG catalog.xml URL and
extracts the "dap" service files, and returns them as a character vector
}
\seealso{
Other imerg: 
\code{\link{construct_gpm_base_url}()},
\code{\link{construct_gpm_product}()},
\code{\link{get_closest_url}()},
\code{\link{get_imerg}()}
}
\concept{imerg}
