% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_imerg_v2}
\alias{get_imerg_v2}
\title{Download GPM IMERG data (version 2)
This version should end up being the final version. More reliable than the original version.}
\usage{
get_imerg_v2(datetime_utc, lon_obs, lat_obs, version = 6)
}
\arguments{
\item{datetime_utc}{Observation time in UTC format YYYY-MM-DD HH:MM:SS. Default is NULL.}

\item{lon_obs}{numeric, Longitude in decimal degrees. Default is NULL.}

\item{lat_obs}{numeric, Latitude in decimal degrees. Default is NULL.}
}
\value{
a dataframe of GPM data for each observation
}
\description{
Download GPM IMERG data (version 2)
This version should end up being the final version. More reliable than the original version.
}
\examples{
\dontrun{
datetime_utc = as.POSIXct("2023-01-01 16:00:00", tz = "UTC")
lon = -105
lat = 40
gpm <- get_imerg_v2(datetime_utc, lon_obs = lon, lat_obs = lat)
}
}
