% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_state}
\alias{get_state}
\title{Geolocate location to assign state association}
\usage{
get_state(lon_obs, lat_obs)
}
\arguments{
\item{lon_obs}{numeric, Longitude in decimal degrees. Default is NULL.}

\item{lat_obs}{numeric, Latitude in decimal degrees. Default is NULL.}
}
\value{
State
}
\description{
Geolocate location to assign state association
}
\examples{
\dontrun{
lon = -105
lat = 40
state <- get_state(lon, lat)
}
}
