% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_tz}
\alias{get_tz}
\title{Get the time zone for an observation with a latitude and longitude}
\usage{
get_tz(lon_obs, lat_obs)
}
\arguments{
\item{lon_obs}{Longitude of observation in decimal degrees (°)}

\item{lat_obs}{Latitude of observation in decimal degrees (°)}
}
\value{
The local standard time zone in the format "Etc/GMT+X"
Where X is the offset in hours from GMT
}
\description{
Get the time zone for an observation with a latitude and longitude
}
\examples{
\dontrun{
lon = -120
lat = 40
get_tz(lon, lat)
}
}
