% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_valid_product_version_for_date}
\alias{get_valid_product_version_for_date}
\title{Get the most optimal (or correct) product version for a given date}
\usage{
get_valid_product_version_for_date(product_version, dateTime)
}
\arguments{
\item{product_version}{character product version}

\item{dateTime}{date character}
}
\value{
character
}
\description{
Get the most optimal (or correct) product version for a given date
}
