% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_convert.R
\name{in_to_mm}
\alias{in_to_mm}
\title{Function to convert inches to millimeters}
\usage{
in_to_mm(ppt_in)
}
\arguments{
\item{ppt_in}{Input precipitation (or other depth quantity) in inches}
}
\value{
Output precipitation (or other depth quantity) in millimeters
}
\description{
Function to convert inches to millimeters
}
\examples{
\dontrun{
ppt_mm <- in_to_mm(ppt_in = 0.5)
}
}
