% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_convert.R
\name{k_to_c}
\alias{k_to_c}
\title{Function to convert temperature in Kelvins to Celsius}
\usage{
k_to_c(temp_K)
}
\arguments{
\item{temp_K}{The temperature in degrees Kelvins}
}
\value{
The temperature in degrees Celsius
}
\description{
Function to convert temperature in Kelvins to Celsius
}
\examples{
\dontrun{
tair_c <- k_to_c(temp_K = 280)
}
}
