% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_access.R
\name{madis_bbox_string}
\alias{madis_bbox_string}
\title{Create bounding box using lat/lon for MADIS data retrieval}
\usage{
madis_bbox_string(lat, lon, deg_filter, digits = 6)
}
\arguments{
\item{lat}{End of search window as POSIX-formatted UTC datetime}

\item{lon}{Start of search window as POSIX-formatted UTC datetime}

\item{deg_filter}{Search radius in degrees}

\item{digits}{Number of digits used to format URL string}
}
\value{
MADIS URL-ready string for selecting station data
}
\description{
Create bounding box using lat/lon for MADIS data retrieval
}
