% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{make_tz_table}
\alias{make_tz_table}
\title{Build a table of timezones and offsets when called from get_tz}
\usage{
make_tz_table()
}
\value{
A table of PST, MST, CST, and EST timezones
}
\description{
Build a table of timezones and offsets when called from get_tz
}
