% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_model.R
\name{model_meteo}
\alias{model_meteo}
\title{Model climate data for an ID, at a location/elevation/time}
\usage{
model_meteo(
  id,
  lon_obs,
  lat_obs,
  elevation,
  datetime_utc,
  meteo_df,
  meta_df,
  n_station_thresh = 5
)
}
\arguments{
\item{id}{character or integer}

\item{lon_obs}{numeric, longitude of observation}

\item{lat_obs}{numeric, latitude of observation}

\item{elevation}{numeric}

\item{datetime_utc}{Date or character}

\item{meteo_df}{data.frame}

\item{meta_df}{data.frame}

\item{n_station_thresh}{numeric, number of stations threshold}
}
\value{
data.frame, maybe with a unique ID for joining...
}
\description{
Model climate data for an ID, at a location/elevation/time
}
