% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_access.R
\name{preprocess_meteo}
\alias{preprocess_meteo}
\title{Preprocess met data (add datetime, put in common format)}
\usage{
preprocess_meteo(network, tmp_met)
}
\arguments{
\item{network}{Network of data to process (HADS, LCD, or WCC)}

\item{tmp_met}{Dataframe of meteorological data (from download_meteo_*)}
}
\value{
Processed dataframe of met data
}
\description{
Preprocess met data (add datetime, put in common format)
}
