% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{product_versions}
\alias{product_versions}
\title{Product versions list}
\format{
An object of class \code{list} of length 4.
}
\usage{
product_versions
}
\description{
A dataset of valid product versions.
}
\keyword{datasets}
