% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_qaqc.R
\name{qaqc_comments}
\alias{qaqc_comments}
\title{QAQC comments column}
\usage{
qaqc_comments(data = data)
}
\arguments{
\item{data}{Dataframe of observation data}
}
\value{
Dataframe with a flag on comments 1 = pass; 0 = flagged
}
\description{
QAQC comments column
}
\examples{
\dontrun{
data = qaqc_comments(data)
}
}
