% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_qaqc.R
\name{qaqc_obs}
\alias{qaqc_obs}
\title{QAQC observation data}
\usage{
qaqc_obs(data = data)
}
\value{
dataframe with dupe (Dupe/Pass) and CONUS (NoCONUS, NoData, Pass) flags
}
\description{
QAQC observation data
}
\examples{
\dontrun{
data_qaqc1 = qaqc_obs(data)
}
}
