% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_qc.R
\name{qc_var}
\alias{qc_var}
\title{Quality control variable internal function}
\usage{
qc_var(var, limit_min, limit_max, sd_threshold)
}
\arguments{
\item{var}{numeric vector}

\item{limit_min}{numeric}

\item{limit_max}{numeric}

\item{sd_threshold}{numeric}
}
\value{
numeric vector that has been quality control checked
}
\description{
Quality control variable internal function
}
