% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_select.R
\name{select_meteo}
\alias{select_meteo}
\title{Function to pare down meteorological data to those closest in time}
\usage{
select_meteo(df, datetime_obs)
}
\arguments{
\item{df}{data.frame}

\item{datetime_obs}{Date or character}
}
\value{
data.frame
}
\description{
Function to pare down meteorological data to those closest in time
}
