% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_access.R
\name{download_meteo_lcd}
\alias{download_meteo_lcd}
\title{Download meteorological data from LCD}
\usage{
download_meteo_lcd(datetime_utc_start, datetime_utc_end, stations)
}
\arguments{
\item{datetime_utc_start}{Start of search window as POSIX-formatted UTC datetime}

\item{datetime_utc_end}{End of search window as POSIX-formatted UTC datetime}

\item{stations}{Dataframe of station metadata (from station_select)}
}
\value{
dataframe of meteorological data
}
\description{
Download meteorological data from LCD
}
