% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metadata_gather.R
\name{gather_meta}
\alias{gather_meta}
\title{Function to gather metadata from meta_all using a vector of station ids}
\usage{
gather_meta(stations)
}
\arguments{
\item{stations}{Vector of station ids}
}
\value{
A dataframe of metadata filter to the station ids in 'stations'
}
\description{
Function to gather metadata from meta_all using a vector of station ids
}
\examples{
stations = c("RRMI4", "INLO2", "JVDU1", "72526604751", "RDKC2", "TABC1", "LSUA2", "BAFO3", "SOAC1", "BNTC1")
gather_meta(stations)
}
