% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_eco_level3}
\alias{get_eco_level3}
\title{Geolocate location to assign ecoregion 3 association}
\usage{
get_eco_level3(lon_obs, lat_obs)
}
\arguments{
\item{lon_obs}{numeric, Longitude in decimal degrees. Default is NULL.}

\item{lat_obs}{numeric, Latitude in decimal degrees. Default is NULL.}
}
\value{
Ecoregion level 3
}
\description{
Geolocate location to assign ecoregion 3 association
}
\examples{
\dontrun{
lon = -105
lat = 40
ecoregion3 <- get_eco_level3(lon, lat)
}
}
