% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_imerg}
\alias{get_imerg}
\title{Get GPM IMERG PLP data for a specified lat/lon and datetime}
\usage{
get_imerg(
  datetime_utc = NULL,
  lon_obs = NULL,
  lat_obs = NULL,
  product_version = NULL,
  verbose = FALSE
)
}
\arguments{
\item{datetime_utc}{Observation time in UTC format YYYY-MM-DD HH:MM:SS. Default is NULL.}

\item{lon_obs}{numeric, Longitude in decimal degrees. Default is NULL.}

\item{lat_obs}{numeric, Latitude in decimal degrees. Default is NULL.}

\item{product_version}{character string of the GPM IMERG product version (e.g. "GPM_3IMERGHHL.06").'
Available versions are "GPM_3IMERGHHL.06", "GPM_3IMERGHHE.06", "GPM_3IMERGHHL.07", and "GPM_3IMERGHH.07". Defaults to IMERG Late Run.}

\item{verbose}{logical, whether to print messages or not. Default is FALSE}
}
\value{
a dataframe of GPM data for each observation
}
\description{
Get GPM IMERG PLP data for a specified lat/lon and datetime
}
\examples{
\dontrun{
datetime_utc = as.POSIXct("2023-01-01 16:00:00", tz = "UTC")
lon = -105
lat = 40
gpm <- get_imerg(datetime_utc, lon_obs = lon, lat_obs = lat)
}
}
\seealso{
Other imerg: 
\code{\link{construct_gpm_base_url}()},
\code{\link{construct_gpm_product}()},
\code{\link{get_closest_url}()},
\code{\link{get_final_urls}()}
}
\concept{imerg}
