% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_preprocess.R
\name{get_imerg}
\alias{get_imerg}
\title{Download GPM IMERG data}
\usage{
get_imerg(datetime_utc, lon_obs, lat_obs, version = 6)
}
\arguments{
\item{datetime_utc}{observation time}

\item{lon_obs}{longitude of observation in decimal degrees (°)}

\item{lat_obs}{latitude of observation in decimal degrees (°)}

\item{version}{IMERG versions: 6 = late run v6 and 7 = final run v7}
}
\value{
a dataframe of GPM data for each observation
}
\description{
Download GPM IMERG data
}
\examples{
\dontrun{
datetime_utc = as.POSIXct("2023-01-01 16:00:00", tz = "UTC")
lon = -105
lat = 40
gpm <- get_imerg(datetime_utc, lon_obs = lon, lat_obs = lat, version = 6)
}
}
