% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meteo_access.R
\name{preprocess_meteo}
\alias{preprocess_meteo}
\title{Preprocess met data (add datetime, put in common format)}
\usage{
preprocess_meteo(network, tmp_met)
}
\arguments{
\item{network}{network of data to process (HADS, LCD, or WCC)}

\item{tmp_met}{dataframe of meteorological data (from download_meteo_*)}
}
\value{
processed dataframe of met data
}
\description{
Preprocess met data (add datetime, put in common format)
}
\examples{
\dontrun{
lon = -105
lat = 40
datetime_start = as.POSIXct("2023-01-01 15:00:00", tz = "UTC")
datetime_end = as.POSIXct("2023-01-01 17:00:00", tz = "UTC")
lcd_stations <- station_select(network = "LCD", lon, lat,
                   deg_filter=2, dist_thresh_m=100000)
lcd_met <- download_meteo_lcd(datetime_start, datetime_end, lcd_stations)
preprocess_meteo("LCD", lcd_met)
}
}
