% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_qaqc.R
\name{qaqc_processed}
\alias{qaqc_processed}
\title{QAQC processed data}
\usage{
qaqc_processed(
  data = data,
  snow_max_tair = 10,
  rain_max_tair = -5,
  rh_min = 30,
  max_avgdist_station = 2e+05,
  max_closest_station = 30000,
  min_n_station = 5,
  pval_max = 0.05
)
}
\arguments{
\item{data}{Dataframe of processed observation data}

\item{snow_max_tair}{Max tair in °C for snow}

\item{rain_max_tair}{Min tair in °C for rain}

\item{rh_min}{Min for RH \%}

\item{max_avgdist_station}{Max average distance (m)}

\item{max_closest_station}{Max nearest station distance (m)}

\item{min_n_station}{Min number of stations within search radius}

\item{pval_max}{Max p-value for the lapse rate calc}
}
\value{
List with 2 dataframes, 1: All flags; 2: Filtered to just incomplete data
}
\description{
QAQC processed data
}
\examples{
\dontrun{
data = data_processed
snow_max_tair = 10,
rain_max_tair = -5,
rh_min = 30,
max_avgdist_station = 2e5,
max_closest_station = 3e4,
min_n_station = 5,
pval_max = 0.05
data_qaqc = qaqc_processed(data = data,
                           snow_max_tair = snow_max_tair,
                           rain_max_tair = rain_max_tair,
                           rh_min = rh_min,
                           max_avgdist_station = max_avgdist_station,
                           max_closest_station = max_closest_station,
                           min_n_station = min_n_station,
                           pval_max = pval_max)
}
}
