% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/citsci_qaqc.R
\name{qaqc_processed}
\alias{qaqc_processed}
\title{QAQC processed data}
\usage{
qaqc_processed(data = data)
}
\value{
list with 2 dataframes, 1: all flags; 2: filtered to the not complete data
}
\description{
QAQC processed data
}
\examples{
\dontrun{
data_qaqc2 = qaqc_processed(data)
}
}
